typing in properties updates document
