+++ src/render/idyll-display/components/author-view.js
@@ -1,26 +1,39 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import PropertyList from './property-list';
+import { getNodeById } from '../utils/';
+import Context from '../../context';
+const AST = require('idyll-ast');
 
 /**
  *
  */
 class AuthorView extends React.PureComponent {
+  static contextType = Context;
+
   constructor(props) {
     super(props);
-    console.log(props.activeComponent);
   }
 
-  updateNodeWithNewProperties(idyllASTNode, propertiesCopy) {
-    console.log(idyllASTNode, propertiesCopy);
+  updateNodeWithNewProperties(idyllASTNode, newPropertyList) {
+    // update node
+    let node = getNodeById(this.context.ast, idyllASTNode.id);
+    const newNode = { ...node, properties: newPropertyList };
+
+    const childrenCopy = AST.getChildren(node);
+    if (newNode.children) {
+      newNode.children = childrenCopy;
+    }
+
+    node.properties = newPropertyList;
+    this.context.setAst(this.context.ast);
   }
 
   render() {
-    const { activeComponent } = this.props;
     const childComponent = (
       <div className='author-view-overlay'>
         <PropertyList
-          node={activeComponent}
+          node={this.context.activeComponent}
           updateNodeWithNewProperties={this.updateNodeWithNewProperties.bind(
             this
           )}
@@ -29,7 +42,7 @@ class AuthorView extends React.PureComponent {
     );
 
     const componentDomNode = document.getElementById(
-      activeComponent.name + '-' + activeComponent.id
+      this.context.activeComponent.name + '-' + this.context.activeComponent.id
     );
 
     if (componentDomNode) {
