+++ src/render/idyll-display/components/property-list.js
@@ -52,96 +52,6 @@ class Component extends React.PureComponent {
     };
   }
 
-  renderExpression(key, prop) {
-    return (
-      <div style={{ display: 'flex', flexDirection: 'row', width: '100%' }}>
-        <input
-          className={'prop-input'}
-          style={{ fontFamily: 'monospace' }}
-          onChange={this.handleUpdateValue(key)}
-          type='text'
-          value={prop.value}
-        ></input>
-        <div
-          className={'prop-type'}
-          onClick={this.handleUpdateType(key, 'variable')}
-          style={{
-            marginLeft: 0,
-            borderRadius: '0 20px 20px 0',
-            background: '#B8E986'
-          }}
-        >
-          {prop.type}
-        </div>
-        <div></div>
-      </div>
-    );
-  }
-  renderValue(key, prop) {
-    return (
-      <div style={{ display: 'flex', flexDirection: 'row', width: '100%' }}>
-        <input
-          className={'prop-input'}
-          onChange={this.handleUpdateValue(key)}
-          type='text'
-          value={prop.value}
-        ></input>
-        <div
-          className={'prop-type'}
-          onClick={this.handleUpdateType(key, 'expression')}
-          style={{
-            marginLeft: 0,
-            borderRadius: '0 20px 20px 0',
-            background: '#4A90E2',
-            color: '#fff'
-          }}
-        >
-          {typeof prop.value}
-        </div>
-        <div>{/* Current Value: {idyllState[prop.value]} */}</div>
-      </div>
-    );
-  }
-  renderVariable(key, prop) {
-    const idyllState = this.context.context.data();
-    return (
-      <div>
-        <div style={{ display: 'flex', flexDirection: 'row', width: '100%' }}>
-          <input
-            className={'prop-input'}
-            style={{ fontFamily: 'monospace' }}
-            onChange={this.handleUpdateValue(key)}
-            type='text'
-            value={prop.value}
-          ></input>
-          <div
-            className={'prop-type'}
-            onClick={this.handleUpdateType(key, 'value')}
-            style={{
-              marginLeft: 0,
-              borderRadius: '0 20px 20px 0',
-              background: '#50E3C2'
-            }}
-          >
-            {prop.type}
-          </div>
-        </div>
-        <div>Current Value: {idyllState[prop.value]}</div>
-      </div>
-    );
-  }
-
-  renderProp(key, prop) {
-    switch (prop.type) {
-      case 'variable':
-        return this.renderVariable(key, prop);
-      case 'value':
-        return this.renderValue(key, prop);
-      case 'expression':
-        return this.renderExpression(key, prop);
-    }
-  }
-
   /**
    *
    * @param {string} propertyName the name of the prop
